pyproject.toml
[project]
name = "apibook-project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.17.1",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.8",
]
